+++ article/nodes/heading/heading.js
@@ -66,9 +66,6 @@ Heading.Prototype = function () {
         return this.level;
     };
 
-    this.getAbstract = function() {
-        return this.document.nodes.abstract_1;
-    }
 };
 
 Heading.Prototype.prototype = Text.prototype;